Use make functions instead of shell
Use make functions instead of the shell for filename manipulations.

Using the shell to manipulate file names and paths can be cumbersome
in a Makefile, especially since you have to wrap it in a call to
the `shell` function, i.e. `$(shell ...)` and escape shell variables
with `$$`.

Make has some nice built-in functions for iteration and manipulating
filenames and paths.  Use those as they make the variable assignment
statements much shorter and easier to read.

Factor the shell logic to extract a filename from a URL into a macro
to make the variable assignment that uses it shorter and easier to
read.

Addresses #2