Make sure that make cleans up intermediate files
Make was only deleting one of the `.zip.conf` curl configuration
files.  I found that this was because the the way we had written
the recipe to generate the curl conf files, all conf files were
created when the rule was run for the first target file. While
this is nice and efficient, it causes subsequent conf files to
not be recorded by make and make won't automatically delete them
even if we specify all the files with the `.INTERMEDIATE` target.

Rewrite our recipe so it only generates the target file on each
run. Specify bash as our shell to make it easier to write this
recipe.

Also refactor the recipe so it doesn't repeat the logic of
determining conf file filenames from URLs.  Using bash and its
support for arrays helps here.

Addresses #2